before change

+++ app/Http/Controllers/TasksController.php
@@ -14,11 +14,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-         $tasks = Task::all();
+         $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $microposts = $user->microposts()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'microposts' => $microposts,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
     }
 
     /**
@@ -47,12 +55,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+        ]);
 
-        return redirect('/');
+        return redirect()->back();
     }
 
     /**
@@ -63,11 +70,17 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
+        $user = User::find($id);
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        $data = [
+            'user' => $user,
+            'tasks' => $tasks,
+        ];
+
+        $data += $this->counts($user);
+
+        return view('users.show', $data);
     }
 
     /**
@@ -95,7 +108,7 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
          $this->validate($request, [
-            'status' => 'required|max:10',
+            'title' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
         
@@ -115,9 +128,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back();
     }
 }
